+++ src/room/PCTransport.ts
@@ -1,5 +1,13 @@
 import { debounce } from 'ts-debounce';
+import { MediaDescription, parse, write } from 'sdp-transform';
 import log from '../logger';
+
+/** @internal */
+interface TrackBitrateInfo {
+  sid: string;
+  codec: string;
+  maxbr: number;
+}
 
 /** @internal */
 export default class PCTransport {
@@ -10,6 +18,8 @@ export default class PCTransport {
   restartingIce: boolean = false;
 
   renegotiate: boolean = false;
+
+  trackBitrates: TrackBitrateInfo[] = [];
 
   onOffer?: (offer: RTCSessionDescriptionInit) => void;
 
@@ -78,11 +88,114 @@ export default class PCTransport {
     // actually negotiate
     log.debug('starting to negotiate');
     const offer = await this.pc.createOffer(options);
+
+    const sdpParsed = parse(offer.sdp ?? '');
+    sdpParsed.media.forEach((media) => {
+      if (media.type === 'audio') {
+        ensureAudioNack(media);
+      } else if (media.type === 'video') {
+        // mung sdp for codec bitrate setting that can't apply by sendEncoding
+        this.trackBitrates.some((trackbr): boolean => {
+          if (!media.msid || !media.msid.includes(trackbr.sid)) {
+            return false;
+          }
+
+          let codecPayload = 0;
+          media.rtp.some((rtp): boolean => {
+            if (rtp.codec.toUpperCase() === trackbr.codec.toUpperCase()) {
+              codecPayload = rtp.payload;
+              return true;
+            }
+            return false;
+          });
+
+          // add x-google-max-bitrate to fmtp line if not exist
+          if (codecPayload > 0) {
+            if (
+              !media.fmtp.some((fmtp): boolean => {
+                if (fmtp.payload === codecPayload) {
+                  if (!fmtp.config.includes('x-google-max-bitrate')) {
+                    fmtp.config += `;x-google-max-bitrate=${trackbr.maxbr}`;
+                  }
+                  return true;
+                }
+                return false;
+              })
+            ) {
+              media.fmtp.push({
+                payload: codecPayload,
+                config: `x-google-max-bitrate=${trackbr.maxbr}`,
+              });
+            }
+          }
+
+          return true;
+        });
+      }
+    });
+
+    offer.sdp = write(sdpParsed);
+    this.trackBitrates = [];
+
     await this.pc.setLocalDescription(offer);
     this.onOffer(offer);
   }
 
+  async createAndSetAnswer(): Promise<RTCSessionDescriptionInit> {
+    const answer = await this.pc.createAnswer();
+    const sdpParsed = parse(answer.sdp ?? '');
+    sdpParsed.media.forEach((media) => {
+      if (media.type === 'audio') {
+        ensureAudioNack(media);
+      }
+    });
+    answer.sdp = write(sdpParsed);
+    await this.pc.setLocalDescription(answer);
+    return answer;
+  }
+
+  setTrackCodecBitrate(sid: string, codec: string, maxbr: number) {
+    this.trackBitrates.push({
+      sid,
+      codec,
+      maxbr,
+    });
+  }
+
   close() {
     this.pc.close();
+  }
+}
+
+function ensureAudioNack(
+  media: {
+    type: string;
+    port: number;
+    protocol: string;
+    payloads?: string | undefined;
+  } & MediaDescription,
+) {
+  // found opus codec to add nack fb
+  let opusPayload = 0;
+  media.rtp.some((rtp): boolean => {
+    if (rtp.codec === 'opus') {
+      opusPayload = rtp.payload;
+      return true;
+    }
+    return false;
+  });
+
+  // add nack rtcpfb if not exist
+  if (opusPayload > 0) {
+    if (!media.rtcpFb) {
+      media.rtcpFb = [];
+    }
+
+    if (!media.rtcpFb.some((fb) => fb.payload === opusPayload && fb.type === 'nack')) {
+      media.rtcpFb.push({
+        payload: opusPayload,
+        type: 'nack',
+      });
+    }
   }
 }