+++ src/room/participant/RemoteParticipant.ts
@@ -210,6 +210,20 @@ export default class RemoteParticipant extends Participant {
         publication = new RemoteTrackPublication(kind, ti.sid, ti.name);
         publication.updateInfo(ti);
         newTracks.set(ti.sid, publication);
+        const existingTrackOfSource = Array.from(this.tracks.values()).find(
+          (publishedTrack) => publishedTrack.source === publication?.source,
+        );
+        if (existingTrackOfSource) {
+          log.warn(
+            `received a second track publication for ${this.identity} with the same source: ${publication.source}`,
+            {
+              oldTrack: existingTrackOfSource,
+              newTrack: publication,
+              participant: this,
+              participantInfo: info,
+            },
+          );
+        }
         this.addTrackPublication(publication);
       } else {
         publication.updateInfo(ti);